reorganize data store and handling
