+++ lib/storeAndTransferData.js
@@ -1,12 +1,33 @@
 'use strict';
 
-const debug = require('debug')('handleWsConnection');
+const debug = require('debug')('storage');
+
+const concat = require('concat-stream'),
+      uuid = require('uuidv4');
+
+const storages = require('./storages');
+
+const childProcess = require('child_process'),
+      PassThrough = require('stream').PassThrough;
 
 const valuesToWrite = [ 0, 1, null ];
+const DefaultState = 2;
+let switches = [];
+let languageSet = 'de_DE';
 
-// const DefaultState = 2;
+const writeToStorage = function (storage) {
+  const passThrough = new PassThrough();
 
-const handleWsConnection = async function (socket, theStorage, bacnetProcess, switches, languageSet) {
+  storage.put('binaries', passThrough, error => {
+    if (error) {
+      throw new Error(error, error.message);
+    }
+  });
+  passThrough.write(JSON.stringify({ language: languageSet, objects: switches.map(z => ({ id: z.id, deviceId: z.deviceId, objectType: z.objectType, objectId: z.objectId, name: z.name, showItem: z.showItem })) }));
+  passThrough.end();
+};
+
+const handleWsConnection = async function (socket, theStorage, bacnetProcess) {
   socket.send('Connection established');
 
   socket.on('message', message => {
@@ -52,7 +73,7 @@ const handleWsConnection = async function (socket, theStorage, bacnetProcess, sw
 
         debug('handleWsConnection::writeLanguage', languageSet);
 
-        theStorage.write({ languageSet });
+        writeToStorage(theStorage);
         break;
       }
       case 'writeObjects': {
@@ -76,7 +97,7 @@ const handleWsConnection = async function (socket, theStorage, bacnetProcess, sw
         }
 
         if (changed) {
-          theStorage.write({ switches });
+          writeToStorage(theStorage);
         }
         break;
       }
@@ -157,4 +178,88 @@ const handleWsConnection = async function (socket, theStorage, bacnetProcess, sw
   });
 };
 
-module.exports = handleWsConnection;
+const storeAndTransferData = function (options) {
+  if (!options) {
+    throw new Error('Options are missing!');
+  }
+  if (!options.storage) {
+    throw new Error('Storage is missing!');
+  }
+  if (!options.options) {
+    throw new Error('StorageOptions are missing!');
+  }
+
+  const theStorage = new storages[options.storage](options.options);
+
+  const bacnetProcess = childProcess.fork('bacnetProcess.js', [ options.options ], { cwd: __dirname });
+
+  const handleToWSS = socket => handleWsConnection(socket, theStorage, bacnetProcess);
+
+  process.on('exit', () => {
+    bacnetProcess.send({ cmd: 'exit' });
+  });
+
+  theStorage.get('binaries', (err, stream) => {
+    debug('initialize storage: ', options);
+    if (err) {
+      debug(err);
+      switches = [];
+    } else {
+      stream.pipe(concat(data => {
+        const { language, objects } = JSON.parse(data);
+
+        debug('initialize storage: ', objects);
+        debug('initialize storage: ', language);
+        languageSet = language;
+        switches = objects.map(val => ({ id: val.id, deviceId: val.deviceId, objectType: val.objectType, objectId: val.objectId, name: val.name, showItem: val.showItem, state: DefaultState, sticky: true }));
+      }));
+    }
+  });
+
+  bacnetProcess.on('message', message => {
+    if (message.cmd === 'gotCOV') {
+      debug('gotCOV', message.value);
+      const nr = switches.findIndex(xx => (xx.deviceId === message.value.deviceId) && (xx.objectType === message.value.objectType) && (xx.objectId === message.value.objectId));
+
+      if (nr !== -1) {
+        if (switches[nr].state !== message.value.value) {
+          switches[nr].state = message.value.value;
+          switches[nr].sticky = true;
+        }
+      }
+    }
+    if (message.cmd === 'gotName') {
+      debug('gotName', message.value);
+      const nr = switches.findIndex(xx => (xx.deviceId === message.value.deviceId) && (xx.objectType === message.value.objectType) && (xx.objectId === message.value.objectId));
+
+      if (nr !== -1) {
+        if (switches[nr].name !== message.value.value) {
+          switches[nr].name = message.value.value;
+          switches[nr].sticky = true;
+          writeToStorage(theStorage);
+        }
+      }
+    }
+    if (message.cmd === 'gotSched') {
+      debug('gotSched', message.value);
+      const nr = switches.findIndex(xx => (xx.deviceId === message.value.deviceId) && (xx.objectType === message.value.objectType) && (xx.objectId === message.value.objectId));
+
+      if (nr !== -1) {
+        switches[nr].weeklySchedule = message.value.value;
+      }
+    }
+    if (message.cmd === 'gotObject') {
+      debug('gotObject', message.value);
+      const nr = switches.findIndex(xx => (xx.deviceId === message.value.deviceId) && (xx.objectType === message.value.objectType) && (xx.objectId === message.value.objectId));
+
+      if (nr === -1) {
+        switches.push({ id: uuid(), deviceId: message.value.deviceId, objectType: message.value.objectType, objectId: message.value.objectId, name: message.value.name, showItem: false, state: DefaultState, sticky: false });
+        writeToStorage(theStorage);
+      }
+    }
+  });
+
+  return handleToWSS;
+};
+
+module.exports = storeAndTransferData;
